完成了方向志愿录取算法，按照第一轮方向前导课程成绩分方向录取，剩下学生按照指定专业基础课成绩排名先后优先录取。

+++ DirectionRegistration.Web/Models/ScoreInfoViewModel.cs
@@ -8,6 +8,6 @@ namespace DirectionRegistration.Models
     public class ScoreInfoViewModel
     {
         public string ScoreName { get; set; }
-        public float ScoreValue { get; set; }
+        public double ScoreValue { get; set; }
     }
 }
